Better parsing for contract name.
1. The solc_wrapper interface returns the latest contract bytecode by
default (since a single solidity file can have multiple contracts), to
do so we need to parse the source file and get the contract/library
names in the order of appearence. To correctly handle this scenario we
need to ignore the contract/library keywords that appear in comments and
also to consider the possibility of bad formatting.

2. The contract code can make use of libraries, since the address is the
returned bytecode might have unresolved symbols, to help the task of
managing symbols a few functions were added, also we try to decode the
bytecode from hexadecimal by default, but if there is an unresolved
symbol it will fail and we just fall back to the original hexadecimal
representation (this might be a usage trap but required for backwards
compability).