solidity_names() recognizes interface
Based directly on top of pyethereum v1.6.1

This is a hack. Solidity 0.4.11 introduces the interface keyword and
breaks the solidity_names() function. This simply tweaks the function
so that it also treats interfaces like contracts and does not break.

I don't believe pyethereum should attempt to parse solidity files on its
own as compatibility issues like this are almost certain to arise also in the
future. For more long-term a different solution should be found where
pyethereum queries solidity itself and does not parse anything on its own.